Add files via upload
In this activity, Alfredo, Miguel and Juan Pablo implemented a group of fake objects that represents the final system. A fake object is an object that has working implementations but is not the same as the production one. Usually, they take some shortcuts and have a simplified version of production code. It conforms to the mandatory contributor interface and can be passed in its place.